added a few tests and fixed some formatting

+++ RecipeBook.java
@@ -235,7 +235,7 @@ public class RecipeBook
         }
         if(desiredYield >= 1)
         {
-            System.out.println("--Ingredient List for " + desiredYield + " people.--");
+            System.out.println("--Ingredient List for " + (int) desiredYield + " people.--");
             for(RecipeIngredient ingredient : currentRecipe.getConvertedIngredients(desiredYield))
             {
                 System.out.println(ingredient.getRecipeIngredient());
